RSICV 32I Architecture Quartus Project
RISC-V is an innovative instruction-set architecture (ISA) that was initially developed to support research and education in computer architecture at UC Berkeley. Its design aspirations have since expanded, aiming to become a standard, free, and open architecture for industry implementations. Characterized by its flexibility and generality, RISC-V is not tailored to specific microarchitectural styles or
technologies, making it suitable for a wide range of hardware implementations, from custom chips to
multicore processors. In this project, a part of RISC-V ISA will be implemented. RISC-V consists of base ISA and extensions. In this project, the Unprivileged RV32I Base Integer
Instruction Set will be implemented. Additionally, one new instruction will be added as an extension.
FENCE, ECALL, EBREAK, and HINT instructions will not be implemented, as they are irrelevant. The extra instruction is XORID. It will take the xor of rs1 with an embedded constant and write the
result to rd.